+++ scripts/bump-version.sh
@@ -9,6 +9,6 @@ NEW_VERSION="${1}"
 echo "Bumping version: ${NEW_VERSION}"
 perl -pi -e "s/^version = \".*?\"/version = \"$NEW_VERSION\"/" Cargo.toml
 perl -pi -e "s/^version = \".*?\"/version = \"$NEW_VERSION\"/" minijinja/Cargo.toml
-perl -pi -e "s/^(insta.*?)version = \".*?\"/\$1version = \"=$NEW_VERSION\"/" minijinja-autoreload/Cargo.toml
+perl -pi -e "s/^(minijinja.*?)version = \".*?\"/\$1version = \"$NEW_VERSION\"/" minijinja-autoreload/Cargo.toml
 
-cd cargo-insta; cargo check
+cargo check --all